requirements.txt
asgiref==3.4.1
autopep8==1.5.7
bcrypt==3.2.0
boto3==1.17.104
botocore==1.20.104
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.65.1
h11==0.12.0
idna==2.10
jmespath==0.10.0
motor==2.4.0
passlib==1.7.4
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.11.4
python-dateutil==2.8.1
python-decouple==3.4
python-multipart==0.0.5
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
